Moved webform related functions to lib directory
